refactor: extract video generation submission service (#5707)
