+++ server/lib/apiRouteCatalog.generated.json
@@ -22306,7 +22306,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 940
+          "line": 930
         }
       ]
     },
@@ -22317,7 +22317,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 1358
+          "line": 1060
         }
       ]
     },
@@ -22328,7 +22328,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 1334
+          "line": 1036
         }
       ]
     },
@@ -22339,7 +22339,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 1363
+          "line": 1065
         }
       ]
     },
@@ -22350,7 +22350,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 1392
+          "line": 1094
         }
       ]
     },
@@ -22361,7 +22361,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 1440
+          "line": 1142
         }
       ]
     },
@@ -22372,7 +22372,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 1412
+          "line": 1114
         }
       ]
     },
@@ -22383,7 +22383,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 1448
+          "line": 1150
         }
       ]
     },
@@ -22394,7 +22394,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 1444
+          "line": 1146
         }
       ]
     },
@@ -22405,7 +22405,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 819
+          "line": 809
         }
       ]
     },
@@ -22416,7 +22416,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 838
+          "line": 828
         }
       ]
     },
@@ -22427,7 +22427,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 1465
+          "line": 1167
         }
       ]
     },
@@ -22438,7 +22438,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 457
+          "line": 447
         }
       ]
     },
@@ -22449,7 +22449,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 461
+          "line": 451
         }
       ]
     },
@@ -22460,7 +22460,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 529
+          "line": 519
         }
       ]
     },
@@ -22471,7 +22471,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 782
+          "line": 772
         }
       ]
     },
@@ -22482,7 +22482,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 767
+          "line": 757
         }
       ]
     },
@@ -22493,7 +22493,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 670
+          "line": 660
         }
       ]
     },
@@ -22504,7 +22504,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 750
+          "line": 740
         }
       ]
     },
@@ -22515,7 +22515,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 499
+          "line": 489
         }
       ]
     },
@@ -22526,7 +22526,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 501
+          "line": 491
         }
       ]
     },
@@ -22537,7 +22537,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 495
+          "line": 485
         }
       ]
     },
@@ -22548,7 +22548,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 407
+          "line": 397
         }
       ]
     },
@@ -22559,7 +22559,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 1482
+          "line": 1184
         }
       ]
     },
@@ -22570,7 +22570,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 927
+          "line": 917
         }
       ]
     },
@@ -22581,7 +22581,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 913
+          "line": 903
         }
       ]
     },
@@ -22592,7 +22592,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 883
+          "line": 873
         }
       ]
     },
@@ -22603,7 +22603,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 893
+          "line": 883
         }
       ]
     },
@@ -22614,7 +22614,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 1433
+          "line": 1135
         }
       ]
     },
@@ -22625,7 +22625,7 @@
       "sources": [
         {
           "source": "server/routes/videoGen.js",
-          "line": 1471
+          "line": 1173
         }
       ]
     },
